up
git-svn-id: https://svn.r-project.org/R-dev-web/trunk@4171 c52295ea-58df-0310-926a-d16021944841

+++ CRAN/QA/BDR/gannet/packages/tests-gcc10/summary2.R
@@ -5,7 +5,7 @@ for(f in files) {
    d <- file.path(dest, f)
    if(any(grepl("multiple definition of", readLines(f), useBytes = TRUE)))
 	file.copy(f, dest, overwrite = TRUE, copy.date = TRUE)
-   else if (file.exists(d) && 
+   else if (file.exists(d) && !file.exists(sub("log", "out", d)) &&
 	    any(grepl("multiple definition of", readLines(d), useBytes = TRUE)))
        file.remove(d)
 }
